Adds Enemy damage timer

+++ DSI-Gameloft/Assets/Scripts/_PureLogicScripts/Bullet.cs
@@ -4,14 +4,14 @@ using System.Collections;
 [System.Serializable]
 public class Bullet {
     #region Members
-    public int m_Power;
+    public float m_Power;
     public int m_Speed;
     #endregion
 
     public Bullet ()
-        : this (1, 20) { }
+        : this (1.0f, 20) { }
 
-    public Bullet (int power, int speed) {
+    public Bullet (float power, int speed) {
         m_Power = power;
         m_Speed = speed;
     }
